moved thread handler
